+++ routes/web.php
@@ -3,5 +3,9 @@
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
-    return view('welcome');
+    return redirect('/chart');
+});
+
+Route::get('/chart', function () {
+    return view('chart');
 });